Change header-name 'parent_id' to 'parent-id'
When deploying the Application using nginx, the header 'parent_id' would
not get forwarded by nginx.
Apparently this is due to the header-name containing the char '_'.

Although one could supposedly add a rule to the nginx config to allow
for forwarding of this type of header, I figure that it is easier to
just change this header-name to be more inline with other headers.